minimize API change
